requirements.txt
boto3==1.38.8
botocore==1.38.8
fastapi==0.115.8
mysql-connector-python==9.2.0
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20